update file handling
